# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "又东风小程序后台 Api 文档",
        "description": "又东风小程序后台 Api 文档",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/admin-api/admin",
            "description": "普通接口地址"
        },
        {
            "url": "/admin-api",
            "description": "通用接口地址"
        }
    ],
    "paths": {
...
        "/keyword/{keyword}": {
            "get": {
                "tags": [
                    "热搜"
                ],
                "summary": "返回热搜词",
                "description": "根据ID查询热搜词",
                "operationId": "App\\Http\\Controllers\\Admin\\SearchHotKeywordsController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/keyword_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "delete": {
                "tags": [
                    "热搜"
                ],
                "summary": "返回删除状态",
                "description": "根据ID删除热搜",
                "operationId": "App\\Http\\Controllers\\Admin\\SearchHotKeywordsController::destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/keyword_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            }
        },
...
        "/ship/{ship}": {
            "get": {
                "tags": [
                    "配送方式"
                ],
                "summary": "通过ID查询配送方式",
                "description": "返回查询到的配送方式",
                "operationId": "App\\Http\\Controllers\\Admin\\ShipController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ship_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "delete": {
                "tags": [
                    "配送方式"
                ],
                "summary": "通过ID删除配送方式",
                "description": "返回是否删除成功",
                "operationId": "App\\Http\\Controllers\\Admin\\ShipController::destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ship_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "patch": {
                "tags": [
                    "配送方式"
                ],
                "summary": "通过ID更新配送方式",
                "description": "返回更新后的配送方式",
                "operationId": "App\\Http\\Controllers\\Admin\\ShipController::update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/ship_in_path_required"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/ship_in_body"
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            }
        },
        "/spec": {
            "get": {
                "tags": [
                    "属性"
                ],
                "summary": "获取属性列表",
                "description": "返回属性列表",
                "operationId": "App\\Http\\Controllers\\Admin\\SpecController::index",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/current_page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "name": "condition",
                        "in": "query",
                        "description": "类型名称",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "属性查询成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/responses/spec"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "post": {
                "tags": [
                    "属性"
                ],
                "summary": "返回保存属性详情",
                "description": "保存属性",
                "operationId": "App\\Http\\Controllers\\Admin\\SpecController::store",
                "requestBody": {
                    "$ref": "#/components/requestBodies/spec_in_body"
                },
                "responses": {
                    "200": {
                        "description": "属性保存成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            }
        },
        "/spec/{spec}": {
            "get": {
                "tags": [
                    "属性"
                ],
                "summary": "返回ID所属属性",
                "description": "通过ID查询属性",
                "operationId": "App\\Http\\Controllers\\Admin\\SpecController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/spec_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/responses/spec"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "delete": {
                "tags": [
                    "属性"
                ],
                "summary": "返回状态",
                "description": "删除属性",
                "operationId": "App\\Http\\Controllers\\Admin\\SpecController::destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/spec_in_path_required"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            },
            "patch": {
                "tags": [
                    "属性"
                ],
                "summary": "返回更新的属性详情",
                "description": "更新属性",
                "operationId": "App\\Http\\Controllers\\Admin\\SpecController::update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/spec_in_path_required"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/spec_in_body"
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/responses/spec"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "用户"
                ],
                "summary": "获取用户列表",
                "description": "返回用户列表",
                "operationId": "App\\Http\\Controllers\\Admin\\UserController::index",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/current_page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "手机号码",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sex",
                        "in": "query",
                        "description": "性别[1:男 2:女]",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nickname",
                        "in": "query",
                        "description": "昵称",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "状态[1:正常, 2:禁用]",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pid_mobile",
                        "in": "query",
                        "description": "推荐人手机号码",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户查询成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/responses/user"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "页面未找到"
                    }
                }
            }
        },
        "/user/status/{user}": {
            "patch": {
                "tags": [
                    "用户"
                ],
                "summary": "修改用户状态",
                "description": "修改用户状态",
                "operationId": "App\\Http\\Controllers\\Admin\\UserController::status",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "用户ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/user_in_body"
                },
                "responses": {
                    "200": {
                        "description": "状态修改成功"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Ship": {
                "title": "配送方式模板",
                "description": "配送方式模板",
                "properties": {
                    "name": {
                        "description": "配送方式名称",
                        "format": "string",
                        "example": "配送方式1"
                    },
                    "type": {
                        "description": "计算方式[1:按重量 2:按件数]",
                        "format": "integer",
                        "example": 1
                    },
                    "has_cod": {
                        "description": "是否货到付款[1:不是 2:是]",
                        "format": "integer",
                        "example": 1
                    },
                    "firstunit": {
                        "description": "首重(单位:克)",
                        "format": "integer",
                        "example": 500
                    },
                    "continueunit": {
                        "description": "续重(单位:克)",
                        "format": "integer",
                        "example": 500
                    },
                    "def_area_fee": {
                        "description": "按地区设置配送费用是否启用默认配送费用[1:启用 2:不启用]",
                        "format": "integer",
                        "example": 1
                    },
                    "area_type": {
                        "description": "地区类型[1:全部地区 2:指定地区]",
                        "format": "integer",
                        "example": 1
                    },
                    "firstunit_price": {
                        "description": "首重费用",
                        "format": "float",
                        "example": 10
                    },
                    "continueunit_price": {
                        "description": "续重费用",
                        "format": "float",
                        "example": 10
                    },
                    "logi_name": {
                        "description": "物流公司名称",
                        "format": "string",
                        "example": "顺丰速运"
                    },
                    "logi_code": {
                        "description": "物流公司编码",
                        "format": "string",
                        "example": "顺丰速运"
                    },
                    "is_def": {
                        "description": "是否默认[1:默认 2:不默认]",
                        "format": "integer",
                        "example": 2
                    },
                    "sort": {
                        "description": "排序",
                        "format": "integer",
                        "example": 100
                    },
                    "status": {
                        "description": "状态[1:正常 2:禁用]",
                        "format": "integer",
                        "example": 1
                    },
                    "free_postage": {
                        "description": "是否包邮[1:包邮 2:不包邮]",
                        "format": "integer",
                        "example": 2
                    },
                    "goodsmoney": {
                        "description": "满多少免运费",
                        "format": "float",
                        "example": 200
                    },
                    "area_fee": {
                        "description": "地区配送费用",
                        "format": "string",
                        "example": {
                            "area_id": [
                                10005,
                                2005,
                                25576
                            ],
                            "firstunit": 500,
                            "continueunit": 500,
                            "firstunit_price": 12,
                            "continueunit_price": 8
                        }
                    }
                },
                "type": "object"
            },
            "Spec": {
                "title": "属性模型",
                "description": "属性模型",
                "properties": {
                    "name": {
                        "type": "string",
                        "format": "string",
                        "example": "内存"
                    },
                    "sort": {
                        "type": "integer",
                        "format": "integer",
                        "example": 100
                    },
                    "details": {
                        "type": "string",
                        "format": "string",
                        "nullable": true
                    },
                    "values": {
                        "description": "@OA\\Property(format=\"string\",example={34,35})",
                        "format": "string",
                        "example": [
                            34,
                            35
                        ]
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "用户模型",
                "description": "用户模型",
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/user_status"
                    }
                },
                "type": "object"
            },
            "user_status": {
                "description": "用户状态[1:正常,2:黑名单]",
                "type": "integer",
                "default": 1,
                "enum": [
                    1,
                    2
                ]
            }
        },
        "responses": {
...
            "spec": {
                "description": "属性信息",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Spec"
                        }
                    }
                }
            },
            "user": {
                "description": "返回用户信息",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                }
            }
        },
        "parameters": {
            "current_page": {
                "name": "current_page",
                "in": "path",
                "description": "当前页",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64",
                    "default": 1
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "path",
                "description": "每页显示数量",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64",
                    "default": 10
                }
            },
            "keyword_in_path_required": {
                "name": "keyword",
                "in": "path",
                "description": "热搜ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "ship_in_path_required": {
                "name": "ship",
                "in": "path",
                "description": "配送方式ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "spec_in_path_required": {
                "name": "spec",
                "in": "path",
                "description": "属性ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "requestBodies": {
            "ship_in_body": {
                "description": "配送方式表单数据",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Ship"
                        }
                    }
                }
            },
            "spec_in_body": {
                "description": "属性表单数据",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Spec"
                        }
                    }
                }
            },
            "user_in_body": {
                "description": "用户表单数据",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                }
            }
        }
    }
}